+++ db/migrate/20260101000011_add_next_id_seq_to_game_sessions.rb
@@ -1,0 +1,5 @@
+class AddNextIdSeqToGameSessions < ActiveRecord::Migration[7.1]
+  def change
+    add_column :game_sessions, :next_id_seq, :integer, null: false, default: 1
+  end
+end